When a comment is added to a sentence with a constituency tree, parse that comment into the tree object
This involves refactoring the StanzaObject, since it is used in both doc and parse_tree
